chore: rename test-fixtures/page to not match globbing pattern of packages
