+++ packages/test-page.html
@@ -1,8 +1,8 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Core Components Test Environment</title>
     <script src="core-datepicker/core-datepicker.min.js"></script>
     <script src="core-dialog/core-dialog.min.js"></script>
@@ -11,8 +11,7 @@
     <script src="core-suggest/core-suggest.min.js"></script>
     <script src="core-tabs/core-tabs.min.js"></script>
     <script src="core-toggle/core-toggle.min.js"></script>
-    <link rel="shortcut icon" href="about:blank">
+    <link rel="shortcut icon" href="about:blank" />
   </head>
-  <body>
-  </body>
+  <body></body>
 </html>
